ref/requirements.txt
azure-ai-projects
azure-identity
openai

pyodbc
python-dotenv

# Agent Framework (for multi-agent workflow)
agent-framework
agent-framework-azure-ai
agent-framework-devui